+++ cpp/open3d/visualization/gui/UIImage.cpp
@@ -32,9 +32,9 @@
 #include "open3d/geometry/Image.h"
 #include "open3d/io/ImageIO.h"
 #include "open3d/visualization/gui/Theme.h"
-#include "open3d/visualization/rendering/Filament/FilamentEngine.h"
-#include "open3d/visualization/rendering/Filament/FilamentResourceManager.h"
 #include "open3d/visualization/rendering/Renderer.h"
+#include "open3d/visualization/rendering/filament/FilamentEngine.h"
+#include "open3d/visualization/rendering/filament/FilamentResourceManager.h"
 
 namespace open3d {
 namespace visualization {
