+++ cpp/open3d/visualization/rendering/filament/FilamentResourceManager.h
